Move FreeAdCredit to src/components folder.
